requirements.txt
transformers==4.31.0
torch==2.0.1
datasets==2.14.4
rouge-score==0.1.2
nltk==3.8.1
pyyaml==6.0.1
tqdm==4.65.0
matplotlib==3.7.2
